remove some comments, remove now unused _schedule_finalizers(), move tests to a pytester test in testing/python/fixtures.py
